requirements.txt
openai
yfinance
python-dotenv
requests
